Support for opening PED format files. (#41)
This commit allows PED format files to be opened by libplinkio_open_txt().
To use it in C, LIBPLINKIO_EXPERIMENTAL must be defined before including "plinkio/plinkio.h".
For use in Python, it can be used by calling, for example, plinkfile.open(txt_data_prefix, is_txt=True, experimental_mode=True).

+++ tests/plinkio_test.c
@@ -12,7 +12,7 @@
 int
 main(int argc, char *argv[])
 {
-    int num_rows;
+    size_t num_rows;
     snp_t *snp_buffer;
     struct pio_file_t plink_file;
